Add ability to pass options (like format)
One downside is that browsers that support woff2 also support
unicode-range. Google Fonts returns a lot more font resources using
unicode-range which won't get downloaded by the browser if not needed, but
do get downloaded and converted by font-store. As a result the woff2 json
is much larger than the woff json (for example over 900kB to almost
300kB).

How to deal with that?